PostListActivity and namespace restructure
Created a PostListActivity which will be the activity for reading posts.
Currently PostListFragment makes an entire new request when there is an activity
transition, which isn't ideal. Either the data isn't being cached and we need to
investigate Loader, or the Fragment is dying when it should stay alive.

Cleaned up the class names a bit.