PostFragment with dummy data
Clicking on a post item in the PostItemActivity will now create a PostFragment
populated with dummy strings and replace the detail container.

TODO: Clean up the way the application stores data internally and to set up the
infrastructure that will allow PostItemActivity to read more complicated data
like comment links.