Reddit API and supporting classes
Added a Reddit, Subreddit, and Query class. The popular method on the Reddit
class returns an array of Subreddit objects, making the API easily callable from
an ASyncTask over in the UI code.

Updated the SubredditListFragment to use this new API scheme. This causes a
regression in major functionality where the application hasn't been updated to
use this new API, or the API is not up to feature parity with the old API call
scheme. This is just the first step towards moving towards a clean API.

added reddit class

generate subreddit class

query class and popular subreddits method

popular method returns subreddit array

hooked up subredditlistfragment to new api